Ingestion API - first version with tests
Included in this version is a Flask app using a Redis-backed queue.
Calls to Redis are wrapped by a circuitbreaker to hopefully lessen
the effect of a slow/non-responsive queue from overloading the ingestion
API workers. We are using Celery to queue and process tasks, and as such
the Flask app relies on Celery task code.

Also included here is a requirements.txt generated from a pip freeze
during development. TODO: remove any extraneous requirements from this.