Fixed bug in the list api
If the number of results to be returned are the count value + 1 then we
can drop the last result as it is only read for working out where in the
database to begin the next page's results from. However, in the case
where the number of results is the same as the count value or less, we
should not drop a result off the end.